+++ tests/org/elixir_lang/model/psi/ClauseKindTest.kt
@@ -26,7 +26,7 @@ class ClauseKindTest : PlatformTestCase() {
         val module = myFixture.file.children.filterIsInstance<Call>().single()
         val actual = module.macroChildCallSequence().joinToString("\n") { clause ->
             val functionSymbol = FunctionSymbol.fromDeclaration(clause).single()
-            val atomSymbol = AtomSymbol.fromClause(clause).single()
+            val atomSymbol = AtomSymbol.fromDeclaration(clause).single()
             val type = ElementDescriptionUtil.getElementDescription(clause, UsageViewTypeLocation.INSTANCE)
 
             "${functionSymbol.name}: FunctionSymbol macro=${functionSymbol.macro}, AtomSymbol macro=${atomSymbol.macro}, $type"